+++ lib/td_core/search/index_worker_impl.ex
@@ -30,6 +30,10 @@ defmodule TdCore.Search.IndexWorkerImpl do
     GenServer.cast(index, {:put_embeddings, ids})
   end
 
+  def refresh_links(index, ids) do
+    GenServer.cast(index, {:refresh_links, ids})
+  end
+
   def get_index_workers do
     :td_core
     |> Application.get_env(TdCore.Search.Cluster, [])
@@ -50,7 +54,8 @@ defmodule TdCore.Search.IndexWorkerImpl do
   def consume(events) do
     index_scope = get_index_template_scope()
 
-    Enum.map(events, fn
+    events
+    |> Enum.map(fn
       %{event: "template_updated", scope: scope} ->
         Map.get(index_scope, String.to_atom(scope))
 
@@ -92,6 +97,13 @@ defmodule TdCore.Search.IndexWorkerImpl do
       fn -> Indexer.put_embeddings(index, ids) end,
       fn millis, _ -> Logger.info("#{index} embeddings put in #{millis} ms") end
     )
+
+    {:noreply, index}
+  end
+
+  @impl GenServer
+  def handle_cast({:refresh_links, ids}, index) do
+    :ok = Indexer.refresh_links(index, ids)
 
     {:noreply, index}
   end
@@ -152,9 +164,8 @@ defmodule TdCore.Search.IndexWorkerImpl do
   defp get_index_template_scope do
     :td_core
     |> get_indexes()
-    |> Enum.map(fn {index, resource} ->
+    |> Map.new(fn {index, resource} ->
       {Keyword.get(resource, :template_scope), index}
     end)
-    |> Map.new()
   end
 end