+++ lib/td_core/search/store.ex
@@ -1,0 +1,20 @@
+defmodule TdCore.Search.Store do
+  @moduledoc """
+  Store for Elasticsearch
+  """
+
+  def fetch(:concepts, ids) do
+    {:ok, data} = TdCluster.TdBg.Store.fetch(ids)
+    data
+  end
+
+  def fetch(:quality_controls, ids) do
+    {:ok, data} = TdCluster.TdQx.Store.fetch(quality_control_ids: ids)
+    data
+  end
+
+  def fetch(:structures, ids) do
+    {:ok, data} = TdCluster.TdDd.Store.fetch(ids)
+    data
+  end
+end