+++ CHANGELOG.md
@@ -6,6 +6,18 @@
 
 - [TD-7883] Update td-cache
 
+## [8.2.1] 2026-02-12
+## [8.2.0] 2026-02-12
+
+### Added
+
+- [TD-7861]
+- [TD-7696]
+  -Access dd store to fetch search data
+  - Refresh search data in lm.
+  - Access bg and qx search data
+  - Enable options for building permissions
+
 ## [8.1.0] 2026-02-09
 
 ### Added